Phase 1 (core): fix framework correctness bugs
- parse_expression: interleave postfix parsing with the binary-operator
  loop. Previously postfix ran once after the loop, so trailing binary
  operators after any postfix form were silently dropped
  (f(1) + 2 parsed as f(1); x IN (...) AND y lost the AND clause).
- Arena: run registered destructors for non-trivially-destructible
  objects at reset/destruction (fixes heap leaks from node vector/string
  members); safe move operations (moved-from arena can no longer corrupt
  the destination's chunks); reset() restores the aligned chunk base;
  allocate_array overflow check + trivially-destructible constraint.
- RecursionGuard no longer leaks a depth increment when the limit error
  throws mid-construction.
- Remove virtuals from CRTP bases: token_name is now a compile-time
  customization point (shadow in derived, dispatched via derived()).
- advance() is no longer [[nodiscard]] (its point is the side effect;
  the header previously violated its own annotation).
- Widen line/column tracking from uint16_t to uint32_t everywhere
  (tokens, SourceLocation, ParseError, tokenizer state); sources beyond
  65,535 lines no longer wrap.
- Tokenizer: lex @> and <@ containment operators (were ERROR tokens).
- Delete dead-and-broken code: core lex/tokenizer.h (never executed;
  UB on first EOF token), hash/perfect_hash.h + hash/compile_time.h
  (never instantiated; not actually a perfect hash), SimpleGenerator
  (uninstantiable), DialectRegistry (abstract value type), the IFNDR
  std::hash specialization, PanicModeRecovery placeholder, unused
  visitor/walker machinery, and stale #if 0 example blocks.
- Rewrite LIFETIME.md: previous 'correct pattern' examples double-copied
  the source and put arena pointers in unique_ptr (UB).

577 tests: identical pass/fail set before and after (566 pass; the 11
failures are pre-existing SQL/MIME bugs addressed next).

+++ sql/include/libglot/sql/lex/tokens.h
@@ -245,8 +245,8 @@ struct Token {
     TokenType type;
     uint32_t start;       // Byte offset in source
     uint32_t end;         // Byte offset (exclusive)
-    uint16_t line;        // Line number (1-indexed)
-    uint16_t col;         // Column number (1-indexed)
+    uint32_t line;        // Line number (1-indexed)
+    uint32_t col;         // Column number (1-indexed)
     const char* text;     // Interned text (for identifiers, keywords, literals)
 
     [[nodiscard]] size_t length() const { return end - start; }
